requirements.txt
absl-py==2.3.1
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
audioread==3.1.0
babel==2.17.0
backports.zstd @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_backports.zstd_1767045020/work
bangla==0.0.5
biopython @ file:///Users/runner/miniforge3/conda-bld/biopython_1761734690346/work
blinker==1.9.0
blis==1.3.3
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1764017116075/work
catalogue==2.0.10
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1767500808759/work/certifi
cffi==2.0.0
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1760437218288/work
click @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_click_1764518652/work
cloudpathlib==0.23.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
confection==0.1.5
contourpy @ file:///Users/runner/miniforge3/conda-bld/contourpy_1762525493111/work
coqpit-config==0.2.3
coqui-tts==0.27.3
coqui-tts-trainer==0.3.2
cycler @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_cycler_1764466758/work
cymem==2.0.13
Cython==3.2.4
dataclassy @ file:///home/conda/feedstock_root/build_artifacts/dataclassy_1673711725182/work
dateparser==1.1.8
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
diffusers==0.36.0
docopt==0.6.2
einops==0.8.1
encodec==0.1.1
exceptiongroup==1.3.1
filelock==3.20.3
Flask==3.1.2
fonttools @ file:///Users/runner/miniforge3/conda-bld/fonttools_1765632680641/work
frozenlist==1.8.0
fsspec==2026.1.0
g2pkk==0.1.2
greenlet @ file:///Users/runner/miniforge3/conda-bld/greenlet_1764863753030/work
grpcio==1.76.0
gruut==2.4.0
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.16.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_h2_1756364871/work
hangul-romanize==0.1.0
hdbscan @ file:///Users/runner/miniforge3/conda-bld/hdbscan_1765731501575/work
hf-xet==1.2.0
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1760286409563/work
importlib_metadata==8.7.1
importlib_resources==6.5.2
inflect==7.5.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1765794027227/work
jsonlines==1.2.0
kalpy-kaldi @ file:///Users/runner/miniforge3/conda-bld/kalpy_1765358609985/work
kiwisolver @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_kiwisolver_1762488860/work
kneed @ file:///home/conda/feedstock_root/build_artifacts/kneed_1735347959166/work
ko-speech-tools==0.1.0
lazy_loader @ file:///home/conda/feedstock_root/build_artifacts/lazy-loader_1733636780672/work
librosa @ file:///home/conda/feedstock_root/build_artifacts/librosa_1741797566785/work
llvmlite==0.46.0
Markdown==3.10
markdown-it-py @ file:///home/conda/feedstock_root/build_artifacts/markdown-it-py_1754951200865/work
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl @ file:///home/conda/feedstock_root/build_artifacts/mdurl_1733255585584/work
monotonic-alignment-search==0.2.1
Montreal_Forced_Aligner @ file:///home/conda/feedstock_root/build_artifacts/montreal-forced-aligner_1761464669803/work
more-itertools==10.8.0
mpmath==1.3.0
msgpack @ file:///Users/runner/miniforge3/conda-bld/msgpack-python_1762504096247/work
multidict==6.7.0
munkres==1.1.4
murmurhash==1.0.15
networkx==2.8.8
nltk==3.9.2
num2words==0.5.14
numba @ file:///Users/runner/miniforge3/conda-bld/numba_1765808305328/work
numpy==1.26.4
ollama==0.6.1
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas==1.5.3
pgvector @ file:///home/conda/feedstock_root/build_artifacts/pgvector-python_1745701581043/work
pillow==12.1.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1764950726/work
pooch @ file:///home/conda/feedstock_root/build_artifacts/pooch_1754941678315/work
praatio @ file:///home/conda/feedstock_root/build_artifacts/praatio_1735072589958/work
preshed==3.0.12
propcache==0.4.1
protobuf==6.33.4
psutil==7.2.1
psycopg2 @ file:///Users/runner/miniforge3/conda-bld/psycopg2-split_1727892685316/work
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
pynini @ file:///Users/runner/miniforge3/conda-bld/pynini_1762534239098/work
pynndescent==0.6.0
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pyparsing_1766475444/work
pypinyin==0.55.0
pysbd==0.3.4
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-crfsuite==0.9.12
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
pytz==2025.2
PyYAML @ file:///Users/runner/miniforge3/conda-bld/pyyaml_1758891744044/work
regex==2025.11.3
requests @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_requests_1766926974/work
rich @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_rich_1760026188/work/dist
rich-click @ file:///home/conda/feedstock_root/build_artifacts/rich-click_1766334403852/work
safetensors==0.7.0
scikit-learn @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_scikit-learn_1766550842/work/dist/scikit_learn-1.8.0-cp311-cp311-macosx_11_0_arm64.whl#sha256=e9bfee154265c2c47be943a5c346c467a67d1d1bed3bb034914a57d76df0c1a9
scipy @ file:///Users/runner/miniforge3/conda-bld/scipy-split_1768134621190/work/dist/scipy-1.17.0-cp311-cp311-macosx_11_0_arm64.whl#sha256=52a502742729a62c926526c452039fe840693d4e6406b81606cb74fefc5bba38
seaborn==0.13.2
shellingham==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
smart_open==7.5.0
soundfile @ file:///home/conda/feedstock_root/build_artifacts/pysoundfile_1737836266465/work
soxr @ file:///Users/runner/miniforge3/conda-bld/soxr-python_1762684782108/work
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy @ file:///Users/runner/miniforge3/conda-bld/sqlalchemy_1765351934917/work
srsly==2.5.2
standard-aifc @ file:///Users/runner/miniforge3/conda-bld/standard-aifc_1761318476720/work
standard-sunau @ file:///Users/runner/miniforge3/conda-bld/standard-sunau_1761318525244/work
SudachiDict-core==20251022
SudachiPy==0.6.10
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
thinc==8.3.10
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1741878222898/work
tokenizers==0.22.2
torch==2.9.1
torchaudio==2.9.1
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1735661334605/work
trainer==0.0.36
transformers==4.57.3
typeguard==4.4.4
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzlocal==5.3.1
umap-learn==0.5.11
unicodedata2 @ file:///Users/runner/miniforge3/conda-bld/unicodedata2_1763054903095/work
Unidecode==1.4.0
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1767817748113/work
wasabi==1.1.3
weasel==0.4.3
Werkzeug==3.1.5
wrapt==2.0.1
yarl==1.22.0
zipp @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_zipp_1764460141/work
